Updating CSS HTML Images files and folder
Adding an Icon folder in the images folder, with one icon image.
Added icon, description, keyword, and viewport meta tags in the head.
Removed the inline styling for the bootstrap elements.
copied and pasted everything in the body and put in the gird container div.
Changed out some of the wording  in the heading and paragraph elements.

For CSS I styled a grid element and made other general styling and layout changes.