Add feature to hide timer when it reaches 0
Updated the JavaScript code to hide the timer element when it reaches 0 using the style.display property. The timer element is now hidden instead of being removed from the DOM. Also updated the CSS to change the opacity of the timer element when it's hidden.
The commit also includes some minor updates to the HTML and CSS files to improve readability and organization.